refactor(react-native): replace useCallControls hook with useMediaStreamManagement
