+++ packages/react-native-sdk/src/hooks/index.ts
@@ -1,8 +1,6 @@
-export * from './useCallControls';
 export * from './useLocalVideoStream';
 export * from './useIncallManager';
 export * from './usePermissionRequest';
 export * from './usePermissionNotification';
 export * from './push';
 export * from './useAndroidKeepCallAliveEffect';
-export * from './useCallControls';
